Several small changes to get the PC working without warnings
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@5879 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/lib/krb5/os/ktdefname.c
@@ -39,7 +39,7 @@ krb5_kt_default_name(context, name, namesize)
     cp = getenv("KRB5_KTNAME");
     if (cp) {
 	strncpy(name, cp, namesize);
-	if (strlen(cp) >= namesize)
+	if (strlen(cp) >= (size_t) namesize)
 	    return KRB5_CONFIG_NOTENUFSPACE;
     } else {
 	strncpy(name, krb5_defkeyname, namesize);
